Removed ubuntu-devel
GitHub Actions doesn't seem to support the `ubuntu-devel` runner
anymore. See list of supported runners on https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners.

Conversely, that runner seems to work fine on other repositories, so
either those are about to fail or the problem isn't quite there.
Nonetheless, this solves the issue of successfully implementing
GHActions as a replacement for the defunct TravisCI (as far as the
open-source community is concerned, anyway). We can deal with adding R-devel
to the CI workflow later.